requirements.txt
appnope==0.1.4
asgiref @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_d2wfxoqk8w/croot/asgiref_1724072473040/work
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
boto3==1.38.14
botocore==1.38.14
captcha==0.7.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.2
decorator==5.2.1
defusedxml==0.7.1
Django @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_8c5bdnaut3/croot/django_1731616018764/work
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
mysqlclient==2.2.7
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
openpyxl==3.1.5
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==1.7.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pyzmq==27.0.0
redis==5.2.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
s3transfer==0.12.0
six==1.17.0
soupsieve==2.7
sqlparse @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_delp7zl09t/croot/sqlparse_1734704454213/work
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
